# report: reuse numberFormatters for ~50% performance gains