<details>
<summary> deferring table rendering stuff that we are _deferring_ to another PR </summary>


baseline perf of renderReport.. ~140ms JS plus rendering:
<img width="766" alt="image" src="https://user-images.githubusercontent.com/39191/199629436-0867bf7f-f208-4595-aec8-3956d5ac5de7.png">


after deferring `renderTable`'s impl.. just 70ms of JS before FCP:
<img width="759" alt="image" src="https://user-images.githubusercontent.com/39191/199629191-f63b1920-f218-47fa-b109-7f95565b4b0c.png">

</details>

and then 

reusing formatters brings 120ms of table rendering down to 50ms:
<img width="659" alt="image" src="https://user-images.githubusercontent.com/39191/199629393-069ad676-712c-4416-92d8-c11887d5ff7d.png">


that is renderReport being 50% faster!!!!!!!