@brendankenny 2022-11-03
v8 has only ever cached the simplest formatters (e.g. the default `toLocaleString` without options). They've added a few cases beyond that, but the advice has always been to reuse formatters with complicated options. Existing bug: https://crbug.com/v8/11159